Do not import xla_model.py from test_util.py, as this ends up picking up the xla_model.py from the installed module, instead of the local one. This prevents proper developer testing. New code ends up picking up the local repo version as first choice.
